Replace widget native versions of ColorBar
This removes the widget native (qt/gtk) implementations of the ColorBar
widget and replaces it with a common version based on a ginga widget.
All the functionality is preserved.

Additionally, the Readout class is changed from using a wrapped Label
widget to using a wrapped ginga widget.  This solves a problem with the
Qt label widget limiting/forcing resize of the reference viewer main
screen.
